requirements.txt
flask==3.0.0
flask-cors==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
chromadb==0.4.22
python-dotenv==1.0.0
schedule==1.2.0
lxml==5.1.0
markdownify==0.11.6
apscheduler==3.10.4
